requirements.txt
snowflake-connector-python
pandas
boto3
streamlit
altair
openai
google-genai
